Add fmt library and use it to format and colour headings
This adds the fmt library as a dependency. It then uses fmt to make
the code for printing headings nicer by using `fmt::print()` and it
also makes use of fmt colouring to print the headings in a nice
green colour so that they can be distinguished from other text.